Refactor, fix Polymer 2.x race condition
This refactor keeps the template and API the same but cleans up
the code a bit and moves the responsibility for showing/hiding
the deck selector dropdown to a single private property. This
property could be exposed publicly if we want to allow the
developer to set it.

This also fixes an issue that appeared in Polymer 2 where the
dropdown would toggle closed then immediately toggle opened
when the user tapped a new deck to select it.